\name{applestarbucks}
\docType{data}
\alias{applestarbucks}
\title{Monthly Stock Prices of Apple and Starbucks from Jan 1994 to Apr 2012}
\description{
This data set gives the monthly stock prices of Apple and Starbucks, in US dollars, from Jan 1994 to Apr 2012.
}
\usage{applestarbucks}
\format{A zoo object containing 220 observations.}
\keyword{datasets}