\name{RelValAnalysis-package}
\alias{RelValAnalysis-package}
\alias{RelValAnalysis}
\docType{package}
\title{
The RelValAnalysis Package
}
\description{
Classes and functions for analyzing the performance of portfolios relative to a benchmark.
}
\details{
\tabular{ll}{
Package: \tab RelValAnalysis\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-25\cr
License: \tab GPL-2 | GPL-3 \cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Ting-Kam Leonard Wong

Maintainer: Ting-Kam Leonard Wong <wongting@uw.edu>
}
\references{
Fernholz, E. R. (2002) \emph{Stochastic portfolio theory}. Springer.

Karatzas, I. and R. Fernholz (2009). Stochastic portfolio theory: an overview. \emph{Handbook of numerical analysis 15}, 89-167.

Pal, S. and T.-K. L. Wong (2013). Energy, entropy, and arbitrage. \emph{arXiv preprint arXiv:1308.5376}.

Pal, S. and T.-K. L. Wong (2014). The geometry of relative arbitrage. \emph{arXiv preprint arXiv:1402.3720}.
}
\keyword{package}
